feat: 回拨版本至github.com/pingcap/tidb v1.1.0-beta.0.20240705114331-eaff0fbc576a

+++ tests/realtikvtest/flashbacktest/flashback_test.go
@@ -22,14 +22,14 @@ import (
 	"time"
 
 	"github.com/pingcap/failpoint"
-	ddlutil "github.com/pingcap/tidb/pkg/ddl/util"
-	"github.com/pingcap/tidb/pkg/errno"
-	"github.com/pingcap/tidb/pkg/meta"
-	"github.com/pingcap/tidb/pkg/meta/model"
-	"github.com/pingcap/tidb/pkg/testkit"
-	"github.com/pingcap/tidb/pkg/testkit/testfailpoint"
-	"github.com/pingcap/tidb/pkg/types"
-	"github.com/pingcap/tidb/tests/realtikvtest"
+	"github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl"
+	ddlutil "github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/util"
+	"github.com/ocean2811/tidbeaff0fbc576a/pkg/domain"
+	"github.com/ocean2811/tidbeaff0fbc576a/pkg/errno"
+	"github.com/ocean2811/tidbeaff0fbc576a/pkg/meta"
+	"github.com/ocean2811/tidbeaff0fbc576a/pkg/parser/model"
+	"github.com/ocean2811/tidbeaff0fbc576a/pkg/testkit"
+	"github.com/ocean2811/tidbeaff0fbc576a/tests/realtikvtest"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/tikv/client-go/v2/oracle"
@@ -61,36 +61,38 @@ func MockGC(tk *testkit.TestKit) (string, string, string, func()) {
 }
 
 func TestFlashback(t *testing.T) {
-	store := realtikvtest.CreateMockStoreAndSetup(t)
+	if *realtikvtest.WithRealTiKV {
+		store := realtikvtest.CreateMockStoreAndSetup(t)
 
-	tk := testkit.NewTestKit(t, store)
+		tk := testkit.NewTestKit(t, store)
 
-	timeBeforeDrop, _, safePointSQL, resetGC := MockGC(tk)
-	defer resetGC()
+		timeBeforeDrop, _, safePointSQL, resetGC := MockGC(tk)
+		defer resetGC()
 
-	tk.MustExec(fmt.Sprintf(safePointSQL, timeBeforeDrop))
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, index i(a))")
-	tk.MustExec("insert t values (1), (2), (3)")
+		tk.MustExec(fmt.Sprintf(safePointSQL, timeBeforeDrop))
+		tk.MustExec("use test")
+		tk.MustExec("drop table if exists t")
+		tk.MustExec("create table t(a int, index i(a))")
+		tk.MustExec("insert t values (1), (2), (3)")
 
-	time.Sleep(1 * time.Second)
+		time.Sleep(1 * time.Second)
 
-	ts, err := tk.Session().GetStore().GetOracle().GetTimestamp(context.Background(), &oracle.Option{})
-	require.NoError(t, err)
+		ts, err := tk.Session().GetStore().GetOracle().GetTimestamp(context.Background(), &oracle.Option{})
+		require.NoError(t, err)
 
-	injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
-	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
-		fmt.Sprintf("return(%v)", injectSafeTS)))
+		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
+			fmt.Sprintf("return(%v)", injectSafeTS)))
 
-	tk.MustExec("insert t values (4), (5), (6)")
-	tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec("insert t values (4), (5), (6)")
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
-	tk.MustExec("admin check table t")
-	require.Equal(t, tk.MustQuery("select max(a) from t").Rows()[0][0], "3")
-	require.Equal(t, tk.MustQuery("select max(a) from t use index(i)").Rows()[0][0], "3")
+		tk.MustExec("admin check table t")
+		require.Equal(t, tk.MustQuery("select max(a) from t").Rows()[0][0], "3")
+		require.Equal(t, tk.MustQuery("select max(a) from t use index(i)").Rows()[0][0], "3")
 
-	require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
+	}
 }
 
 func TestPrepareFlashbackFailed(t *testing.T) {
@@ -114,12 +116,12 @@ func TestPrepareFlashbackFailed(t *testing.T) {
 		require.NoError(t, err)
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/mockPrepareMeetsEpochNotMatch", `return(true)`))
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/mockPrepareMeetsEpochNotMatch", `return(true)`))
 
 		tk.MustExec("insert t values (4), (5), (6)")
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		tk.MustExec("admin check table t")
 		require.Equal(t, tk.MustQuery("select max(a) from t").Rows()[0][0], "3")
@@ -130,8 +132,8 @@ func TestPrepareFlashbackFailed(t *testing.T) {
 		require.NoError(t, job.Decode([]byte(jobMeta)))
 		require.Equal(t, job.ErrorCount, int64(0))
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/mockPrepareMeetsEpochNotMatch"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/mockPrepareMeetsEpochNotMatch"))
 	}
 }
 
@@ -163,18 +165,18 @@ func TestFlashbackAddDropIndex(t *testing.T) {
 		require.Greater(t, tk.MustQuery("select count(*) from mysql.gc_delete_range").Rows()[0][0], prevGCCount)
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
 
 		tk.MustExec("insert t values (4), (5), (6)")
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		tk.MustExec("admin check table t")
 		require.Equal(t, tk.MustQuery("select max(a) from t use index(i)").Rows()[0][0], "3")
 		tk.MustGetErrCode("select max(a) from t use index(k)", errno.ErrKeyDoesNotExist)
 		require.Equal(t, tk.MustQuery("select count(*) from mysql.gc_delete_range").Rows()[0][0], prevGCCount)
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
@@ -208,11 +210,11 @@ func TestFlashbackAddDropModifyColumn(t *testing.T) {
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin")
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
 
 		tk.MustExec("insert t values (4, 4), (5, 5), (6, 6)")
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		tk.MustExec("admin check table t")
 		require.Equal(t, tk.MustQuery("show create table t").Rows()[0][1], "CREATE TABLE `t` (\n"+
@@ -222,7 +224,7 @@ func TestFlashbackAddDropModifyColumn(t *testing.T) {
 			") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin")
 		require.Equal(t, tk.MustQuery("select max(b) from t").Rows()[0][0], "3")
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
@@ -261,9 +263,9 @@ func TestFlashbackBasicRenameDropCreateTable(t *testing.T) {
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
 
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		tk.MustExec("admin check table t")
 		require.Equal(t, tk.MustQuery("select max(a) from t").Rows()[0][0], "3")
@@ -271,7 +273,7 @@ func TestFlashbackBasicRenameDropCreateTable(t *testing.T) {
 		require.Equal(t, tk.MustQuery("select max(a) from t1").Rows()[0][0], "6")
 		require.Equal(t, tk.MustQuery("select count(*) from mysql.gc_delete_range").Rows()[0][0], prevGCCount)
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
@@ -299,14 +301,14 @@ func TestFlashbackCreateDropTableWithData(t *testing.T) {
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
 
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		tk.MustExec("admin check table t")
 		require.Equal(t, tk.MustQuery("select count(a) from t").Rows()[0][0], "0")
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
@@ -338,17 +340,17 @@ func TestFlashbackCreateDropSchema(t *testing.T) {
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
 
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
 		tk.MustExec("admin check table test.t")
 		res := tk.MustQuery("select max(a) from test.t").Rows()
 		require.Equal(t, res[0][0], "2")
 		tk.MustGetErrCode("use test1", errno.ErrBadDB)
 		tk.MustGetErrCode("use test2", errno.ErrBadDB)
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
@@ -377,9 +379,9 @@ func TestFlashbackAutoID(t *testing.T) {
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
 
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		tk.MustExec("admin check table t")
 		res = tk.MustQuery("select max(a) from t").Rows()
@@ -388,7 +390,7 @@ func TestFlashbackAutoID(t *testing.T) {
 		res = tk.MustQuery("select max(a) from t").Rows()
 		require.Equal(t, res[0][0], "101")
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
@@ -417,16 +419,15 @@ func TestFlashbackSequence(t *testing.T) {
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
 
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		// flashback schema and skip cached values
 		res = tk.MustQuery("select nextval(seq)").Rows()
 		require.Equal(t, res[0][0], "101")
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
-		tk.MustExec("drop sequence seq")
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
@@ -446,7 +447,7 @@ func TestFlashbackPartitionTable(t *testing.T) {
 			"partition `a_2` values less than (75), " +
 			"partition `a_3` values less than (200))")
 
-		for i := range 100 {
+		for i := 0; i < 100; i++ {
 			tk.MustExec(fmt.Sprintf("insert into t values (%d)", i))
 		}
 
@@ -462,9 +463,9 @@ func TestFlashbackPartitionTable(t *testing.T) {
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
 
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		tk.MustExec("admin check table t")
 		res = tk.MustQuery("select max(a), min(a), count(*) from t").Rows()
@@ -477,7 +478,7 @@ func TestFlashbackPartitionTable(t *testing.T) {
 		require.Equal(t, res[0][1], "-1")
 		require.Equal(t, res[0][2], "102")
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
@@ -503,9 +504,9 @@ func TestFlashbackTmpTable(t *testing.T) {
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
 
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		res := tk.MustQuery("select max(a) from t").Rows()
 		require.Equal(t, res[0][0], "3")
@@ -519,19 +520,21 @@ func TestFlashbackTmpTable(t *testing.T) {
 
 		injectSafeTS = oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
 
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
-		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
+		tk.MustExec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
 
 		tk.MustGetErrCode("select * from t", errno.ErrNoSuchTable)
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
 	}
 }
 
 func TestFlashbackInProcessErrorMsg(t *testing.T) {
 	if *realtikvtest.WithRealTiKV {
-		store := realtikvtest.CreateMockStoreAndSetup(t)
+		store, dom := realtikvtest.CreateMockStoreAndDomainAndSetup(t)
+
+		originHook := dom.DDL().GetHook()
 
 		tk := testkit.NewTestKit(t, store)
 		timeBeforeDrop, _, safePointSQL, resetGC := MockGC(tk)
@@ -551,14 +554,15 @@ func TestFlashbackInProcessErrorMsg(t *testing.T) {
 		tk.MustExec("insert into t values (1), (2), (3)")
 
 		injectSafeTS := oracle.GoTimeToTS(oracle.GetTimeFromTS(ts).Add(100 * time.Second))
-		require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS",
+		require.NoError(t, failpoint.Enable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS",
 			fmt.Sprintf("return(%v)", injectSafeTS)))
 
-		testfailpoint.EnableCall(t, "github.com/pingcap/tidb/pkg/ddl/beforeRunOneJobStep", func(job *model.Job) {
+		hook := newTestCallBack(t, dom)
+		hook.OnJobRunBeforeExported = func(job *model.Job) {
 			if job.Type == model.ActionFlashbackCluster && job.SchemaState == model.StateWriteReorganization {
 				txn, err := store.Begin()
 				assert.NoError(t, err)
-				_, err = meta.NewMutator(txn).ListDatabases()
+				_, err = meta.NewMeta(txn).ListDatabases()
 				errorMsg := err.Error()
 				assert.Contains(t, errorMsg, "is in flashback progress, FlashbackStartTS is ")
 				slices := strings.Split(errorMsg, "is in flashback progress, FlashbackStartTS is ")
@@ -566,10 +570,30 @@ func TestFlashbackInProcessErrorMsg(t *testing.T) {
 				assert.NotEqual(t, slices[1], "0")
 				txn.Rollback()
 			}
-		})
-		tk.Exec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts).Format(types.TimeFSPFormat)))
-		testfailpoint.Disable(t, "github.com/pingcap/tidb/pkg/ddl/beforeRunOneJobStep")
+		}
+		dom.DDL().SetHook(hook)
+		tk.Exec(fmt.Sprintf("flashback cluster to timestamp '%s'", oracle.GetTimeFromTS(ts)))
+		dom.DDL().SetHook(originHook)
 
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/pkg/ddl/injectSafeTS"))
+		require.NoError(t, failpoint.Disable("github.com/ocean2811/tidbeaff0fbc576a/pkg/ddl/injectSafeTS"))
+	}
+}
+
+type testCallback struct {
+	ddl.Callback
+	OnJobRunBeforeExported func(job *model.Job)
+}
+
+func newTestCallBack(t *testing.T, dom *domain.Domain) *testCallback {
+	defHookFactory, err := ddl.GetCustomizedHook("default_hook")
+	require.NoError(t, err)
+	return &testCallback{
+		Callback: defHookFactory(dom),
+	}
+}
+
+func (c *testCallback) OnJobRunBefore(job *model.Job) {
+	if c.OnJobRunBeforeExported != nil {
+		c.OnJobRunBeforeExported(job)
 	}
 }
